Added deserializer and refactored
